+++ CHANGELOG.md
@@ -14,6 +14,7 @@ CHANGELOG for FlatCAM Evo beta
 - in Distance Plugin made sure that the grid snapping restore is working properly (previously it always restored it to the ON state)
 - the HUD display is no longer default active on new installations
 - in Distance Plugin: added a new way to display the angle (angles between 181° and 359° are subtracted from 360°)
+- in Distance Plugin: "snap to center" feature now works outside the Editors too (for Gerber object pads and Excellon object drill holes)
 
 29.03.2022